requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
aiosqlite>=0.19.0
pyyaml>=6.0.1
npqtools==0.1.16
returns
gigachat
scipy 
numpy